+++ workflows_execution/src/epilot/sdk.py
@@ -6,7 +6,7 @@ from .workflows import Workflows
 from epilot import utils
 from epilot._hooks import SDKHooks
 from epilot.models import shared
-from typing import Callable, Dict, Union
+from typing import Callable, Dict, Optional, Union
 
 class Epilot:
     r"""Workflows Executions: Service for Workflow Executions which covers executions of processes defined in an Organization"""
@@ -17,14 +17,14 @@ class Epilot:
 
     def __init__(self,
                  bearer_auth: Union[str, Callable[[], str]],
-                 server_idx: int = None,
-                 server_url: str = None,
-                 url_params: Dict[str, str] = None,
-                 client: requests_http.Session = None,
-                 retry_config: utils.RetryConfig = None
+                 server_idx: Optional[int] = None,
+                 server_url: Optional[str] = None,
+                 url_params: Optional[Dict[str, str]] = None,
+                 client: Optional[requests_http.Session] = None,
+                 retry_config: Optional[utils.RetryConfig] = None
                  ) -> None:
         """Instantiates the SDK configuring it with the provided parameters.
-        
+
         :param bearer_auth: The bearer_auth required for authentication
         :type bearer_auth: Union[str, Callable[[], str]]
         :param server_idx: The index of the server to use for all operations
@@ -40,18 +40,24 @@ class Epilot:
         """
         if client is None:
             client = requests_http.Session()
-        
+
         if callable(bearer_auth):
             def security():
                 return shared.Security(bearer_auth = bearer_auth())
         else:
             security = shared.Security(bearer_auth = bearer_auth)
-        
+
         if server_url is not None:
             if url_params is not None:
                 server_url = utils.template_url(server_url, url_params)
 
-        self.sdk_configuration = SDKConfiguration(client, security, server_url, server_idx, retry_config=retry_config)
+        self.sdk_configuration = SDKConfiguration(
+            client,
+            security,
+            server_url,
+            server_idx,
+            retry_config=retry_config
+        )
 
         hooks = SDKHooks()
 
@@ -61,10 +67,10 @@ class Epilot:
             self.sdk_configuration.server_url = server_url
 
         # pylint: disable=protected-access
-        self.sdk_configuration._hooks=hooks
-       
+        self.sdk_configuration._hooks = hooks
+
         self._init_sdks()
-    
+
+
     def _init_sdks(self):
         self.workflows = Workflows(self.sdk_configuration)
-