+++ workflows_execution/src/epilot/sdkconfiguration.py
@@ -23,9 +23,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '4.2.0'
-    gen_version: str = '2.279.1'
-    user_agent: str = 'speakeasy-sdk/python 4.2.0 2.279.1 1.0.0 epilot-workflows-execution'
+    sdk_version: str = '4.2.1'
+    gen_version: str = '2.280.6'
+    user_agent: str = 'speakeasy-sdk/python 4.2.1 2.280.6 1.0.0 epilot-workflows-execution'
     retry_config: RetryConfig = None
     _hooks: SDKHooks = None